requirements.txt
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
cupy==2.1.0.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
fastrlock==0.3
ftfy==4.4.3
html5lib==0.9999999
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.2.2
numpy==1.13.3
olefile==0.44
pandocfilters==1.4.2
parso==0.1.0
pathlib==1.0.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
Pygments==2.2.0
pynvrtc==8.0
pyparsing==2.2.0
python-dateutil==2.6.1
PyTorch-QRNN==0.2.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
spacy==2.0.5
tensorboard-logger==0.0.4
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.8.2
torch==0.3.0.post4
torchtext==0.2.0
torchvision==0.2.0
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
widgetsnbextension==3.0.8
wrapt==1.10.11
